First Version of expand function
The attached file is my implementation of the expand function with a test int main() to display how it is working-ish. Due to how the function looks at only the size of the vector and doesn't care about indexing, the first element will be at index 0, the second at 1, third at 2 and so on.

For intelligently triggering the usage of the expand function I attempted to make a few bare-bones functions to detect if a particular node is_empty (via if Event == "") and if the tree is_full by if (! (data.back().is_empty())). The function empty in this code is